Refactor API routes into separate blueprint and fix CSRF issues
